+++ src/main/java/com/example/proj2/models/Tipopagamento.java
@@ -12,16 +12,16 @@ import java.math.BigDecimal;
 public class Tipopagamento {
     @Id
     @Column(name = "idtipopagamento", nullable = false, precision = 10)
-    private BigDecimal id;
+    private Integer id;
 
     @Column(name = "descricao", length = 50)
     private String descricao;
 
-    public BigDecimal getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Integer id) {
         this.id = id;
     }
