+++ src/main/java/com/example/proj2/repository/LicencaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface LicencaRepository extends JpaRepository<Licenca, BigDecimal> {
+public interface LicencaRepository extends JpaRepository<Licenca, Integer> {
 }
