+++ src/main/java/com/example/proj2/repository/OrcamentoprojetoRepository.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
-public interface OrcamentoprojetoRepository extends JpaRepository<Orcamentoprojeto, BigDecimal> {
+public interface OrcamentoprojetoRepository extends JpaRepository<Orcamentoprojeto, Integer> {
 
     @Query("SELECT o FROM Orcamentoprojeto o LEFT JOIN FETCH o.projeto p LEFT JOIN FETCH p.gestordeprojeto WHERE o.estado = 'pendente de orçamento'")
     List<Orcamentoprojeto> findAllPendingWithProjetoAndGestordeprojeto();
