+++ src/main/java/com/example/proj2/repository/ProjetostecnicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface ProjetostecnicoRepository extends JpaRepository<Projetostecnico, BigDecimal> {
+public interface ProjetostecnicoRepository extends JpaRepository<Projetostecnico, Integer> {
 }
