+++ src/main/java/com/example/proj2/repository/SolicitacaoprojetoRepository.java
@@ -4,5 +4,5 @@ import com.example.proj2.models.Solicitacaoprojeto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigDecimal;
 
-public interface SolicitacaoprojetoRepository extends JpaRepository<Solicitacaoprojeto, BigDecimal> {
+public interface SolicitacaoprojetoRepository extends JpaRepository<Solicitacaoprojeto, Integer> {
 }
