+++ src/main/java/com/example/proj2/repository/TipopagamentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface TipopagamentoRepository extends JpaRepository<Tipopagamento, BigDecimal> {
+public interface TipopagamentoRepository extends JpaRepository<Tipopagamento, Integer> {
 }
