+++ src/main/java/com/example/proj2/repository/TipoutilizadorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
 
-public interface TipoutilizadorRepository extends JpaRepository<Tipoutilizador, BigDecimal> {
+public interface TipoutilizadorRepository extends JpaRepository<Tipoutilizador, Integer> {
 }
